fix: taosbenchmark code refactor (#599)
* fix: refactor line protocol data generation

* test: change test branch

* test: add brew install pkg-config in macos workflows

* fix: check snprintf return

* fix: taosbenchmark insert refactor

* fix: reduce complexity of sml data generating

* test: don't build jdbc on windows

* fix: check unsigned int generated

* fix: refactor cleanup and print

* fix: refect prepare sml data

* fix: add sml_tags_json_array to accelerate sml json test

* fix: taosbenchmark json data generating refine

* fix: comma adding timing

* fix: coverity scan issues

* fix: use strncat instead of snprintf

* fix: adjust end time position

* fix: avoid random data generation

* fix: remove unused func

* fix: remove refactor macro

* test: add taosbenchmark/csv/insert-json-csv.py

* fix: csv read pos/len bug

* fix: support tdengine v2

* fix: don't repeat

+++ src/benchInsert.c
@@ -1813,9 +1813,7 @@ static int32_t prepareProgressDataSmlJson(
 }
 
 static int32_t prepareProgressDataSmlLine(
-    threadInfo *pThreadInfo,
-    uint64_t tableSeq,
-    char *sampleDataBuf,
+    threadInfo *pThreadInfo, uint64_t tableSeq, char *sampleDataBuf,
     int64_t *timestamp, uint64_t i, char *ttl) {
     // prepareProgressDataSmlLine
     SSuperTable *stbInfo = pThreadInfo->stbInfo;
@@ -1828,11 +1826,9 @@ static int32_t prepareProgressDataSmlLine(
                 pThreadInfo->lines[j],
                 stbInfo->lenOfCols + stbInfo->lenOfTags,
                 "%s %s %" PRId64 "",
-                pThreadInfo
-                ->sml_tags[(int)tableSeq -
-                pThreadInfo->start_table_from],
-                sampleDataBuf +
-                pos * stbInfo->lenOfCols,
+                pThreadInfo->sml_tags[tableSeq
+                    - pThreadInfo->start_table_from],
+                sampleDataBuf + pos * stbInfo->lenOfCols,
                 *timestamp);
         pos++;
         if (pos >= g_arguments->prepared_rand) {
@@ -1851,9 +1847,7 @@ static int32_t prepareProgressDataSmlLine(
 }
 
 static int32_t prepareProgressDataSmlTelnet(
-    threadInfo *pThreadInfo,
-    uint64_t tableSeq,
-    char *sampleDataBuf,
+    threadInfo *pThreadInfo, uint64_t tableSeq, char *sampleDataBuf,
     int64_t *timestamp, uint64_t i, char *ttl) {
     // prepareProgressDataSmlTelnet
     SSuperTable *stbInfo = pThreadInfo->stbInfo;
@@ -1868,10 +1862,9 @@ static int32_t prepareProgressDataSmlTelnet(
                 "%s %" PRId64 " %s %s", stbInfo->stbName,
                 *timestamp,
                 sampleDataBuf
-                + pos * stbInfo->lenOfCols,
-                pThreadInfo
-                ->sml_tags[(int)tableSeq
-                -pThreadInfo->start_table_from]);
+                    + pos * stbInfo->lenOfCols,
+                pThreadInfo->sml_tags[tableSeq
+                    -pThreadInfo->start_table_from]);
         pos++;
         if (pos >= g_arguments->prepared_rand) {
             pos = 0;
